Improved file organization.
