+++ routes/index.js
@@ -3,6 +3,6 @@ const express = require('express');
 let router = express.Router();
 
 // Entry point for the github cards
-router.get('/user/:user', require('./githubcards.js'));
+router.use('/user/', require('./githubcards.js'));
 
 module.exports = router;
